fix `arithmetic-shift' with shifts close to `(- (expt 2 31))' on 64-bit
The bug was caused by an accidental cast of a 64-bit integer to a
32-bit integer.

Closes PR 13923

+++ racket/src/racket/src/number.c
@@ -4120,7 +4120,7 @@ scheme_bitwise_shift(int argc, Scheme_Object *argv[])
 
     if (i > 0) {
       if (shift < 0) {
-	int shft = -shift;
+	intptr_t shft = -shift;
 	if (shft < MAX_SHIFT_EVER) {
 	  i = i >> shft;
 	  return scheme_make_integer(i);
